Change from ushort to uint. Partially working file reader.

+++ ZSMapFormat/FormatStructure.cs
@@ -18,7 +18,7 @@ namespace ZSMapFormat
             public string MapCreator; //Level designer
             public string Version; //Version of map
             public byte MapType; //Type of map (regular, objective, etc.)
-            public ushort AmountOfMaterials; //Amount of materials stored in file
+            public uint AmountOfMaterials; //Amount of materials stored in file
             public char[] MaterialsLocation; //Where the materials start in file
             public List<MaterialStorage> Materials; //List containing materials and material path
             public char[] MapDataLocation; //Where the Map Data starts in the file
